sprintf to snprintf - sprintf is deprecated and should be changed with snprintf. This is   done for all of AliRoot with this PR - Notable things:     - in `./ITS/ShowSPDRecPoints.C` there were cases where the lenth of       the const char objects was too small for the messages that are       streamed into them. I increased the length to ensure memory       safety.     - in `AD/ADrec/AliADQADataMakerRec.cxx` it was `if(mulADC!=0)FillRawsData(kChargeADC_PC,chargeTrigADA/mulADC)` which I changed to `if(mulADC!=0)FillRawsData(kChargeADC_PC,chargeTrigADC/mulADC);` since it seemed correct to divide ADC by ADC and not ADA by ADC

+++ STEER/ESD/AliFMDMap.h
@@ -51,15 +51,15 @@ public:
     /** Base index for inner rings */ 
     kBaseInner = 0, 
     /** Base index for outer rings */
-    kBaseOuter = kNSectorInner * kNStripInner
+    kBaseOuter = (int) kNSectorInner * (int) kNStripInner
   };
   enum { 
     /** Base for FMD1 index */ 
     kFMD1Base = 0, 
     /** Base for FMD2 index */ 
-    kFMD2Base = kNSectorInner * kNStripInner, 
+    kFMD2Base = (int) kNSectorInner * (int) kNStripInner, 
     /** Base for FMD3 index */ 
-    kFMD3Base = (kBaseOuter + kNSectorOuter * kNStripOuter + kFMD2Base)
+    kFMD3Base = ((int) kBaseOuter + (int) kNSectorOuter * (int) kNStripOuter + (int) kFMD2Base)
   };
   /** 
    * Class to do stuff on each element of a map in an efficient
